+++ Engine/Source/Runtime/Engine/Public/VisualLogger/VisualLoggerTypes.h
@@ -396,7 +396,7 @@ FVisualLogShapeElement::FVisualLogShapeElement(const FString& InDescription, con
 inline
 void FVisualLogShapeElement::SetColor(const FColor& InColor)
 {
-	Color = ((InColor.DWColor() >> 30) << 6)	| (((InColor.DWColor() & 0x00ff0000) >> 22) << 4)	| (((InColor.DWColor() & 0x0000ff00) >> 14) << 2)	| ((InColor.DWColor() & 0x000000ff) >> 6);
+	Color = (uint8)(((InColor.DWColor() >> 30) << 6)	| (((InColor.DWColor() & 0x00ff0000) >> 22) << 4)	| (((InColor.DWColor() & 0x0000ff00) >> 14) << 2)	| ((InColor.DWColor() & 0x000000ff) >> 6));
 }
 
 inline
